move helper functions to utils.js
