+++ src/PlayArea/PlayArea.js
@@ -3,7 +3,7 @@ import Keyboard from './Keyboard';
 import PreviousGuesses from './PreviousGuesses';
 import CurrentGuess from './CurrentGuess';
 import './PlayArea.css';
-import statuses from '../statuses';
+import { statuses } from '../utils';
 
 //The main interactive area of the app
 //manages currentGuess and keyStatuses state
@@ -147,37 +147,12 @@ const PlayArea = (props) => {
     bottomOfPlayAreaRef.current.scrollIntoView({ behavior: 'smooth' });
   }, [previousGuesses, wordNum]);
 
-  //assign a className to word-row for styling dependent on wordNum
-  let wordRowClassName = 'word-row';
-  switch(wordNum) {
-    case 0:
-      wordRowClassName += ' first';
-      break;
-    case 1:
-      wordRowClassName += ' second';
-      break;
-    case 2:
-      wordRowClassName += ' third';
-      break;
-    case 3:
-      wordRowClassName += ' fourth';
-      break;
-    case 4:
-      wordRowClassName += ' fifth';
-      break;
-    case 5:
-      wordRowClassName += ' sixth';
-      break;
-    default: 
-      break;
-  }
-
   return (
     <>
       <section id='play-area'>
         <div id='play-area-overflow-scroll'>
-          <PreviousGuesses previousGuesses={previousGuesses} wordNum={wordNum} wordRowClassName={wordRowClassName} />
-          <CurrentGuess currentGuess={currentGuess} wordLength={wordLength} wordRowClassName={wordRowClassName} />
+          <PreviousGuesses previousGuesses={previousGuesses} wordNum={wordNum} />
+          <CurrentGuess currentGuess={currentGuess} wordLength={wordLength} wordNum={wordNum}/>
           <div ref={bottomOfPlayAreaRef} />
         </div>
       </section>
