--- src/statuses.js
@@ -1,4 +1,0 @@
-//the three possible statuses for progressIcons, previousGuess Letters, and Keyboard Keys
-const statuses = ['absent', 'misplaced', 'correct'];
-
-export default statuses;
